binary search added first changes from MVC to MVVM pattern
