+++ DataFromJSONfile/TableViewModelType.swift
@@ -1,0 +1,10 @@
+//
+//  TableViewModelType.swift
+//  DataFromJSONfile
+//
+//  Created by Vasili Staratsitarau on 25/08/2022.
+//
+
+import Foundation
+
+
